:arrows_counterclockwise: Refactor Function module architecture
- Replace complex Backend hierarchy with simple symbol-based backends
- Move JavaScript formatters to Foxtail::Function::JavaScript module
- Unify interface: Proc functions wrap new(locale:, **options).call(value)
- Add format alias for JavaScript formatters
- Fix backend detection by requiring 'locale' gem
- Simplify Function.defaults to return Proc functions for Bundle compatibility
- Remove Backend::Base and related complexity
- Update ExecJS setup in spec_helper with before/after(:suite)
- Clean up RuboCop exclusions for deleted files
- Update all tests for new unified interface

:robot: Generated with [Claude Code](https://claude.ai/code)
